commandHauls: extract haul stats widget to module
