+++ src/Widget/AbstractWidget.js
@@ -1,0 +1,7 @@
+class AbstractWidget {
+    insertAfter($el) {
+        $el.after(this.$el);
+    }
+}
+
+export { AbstractWidget }
